move some protobuf tools out to protobuf_common.go
Methods not tied to traces are now in the common file.
